fix: move P-gate from phase.py into rotation.py since the name 'phase' conflict with S-gate
